Refactored utils into new format.
