+++ tsdat/utils/standards_type.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class StandardsType(str, Enum):
+    tsdat = "tsdat"
+    acdd = "acdd"
+    ioos = "ioos"
